+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/animation/viewport/MapViewportState.kt
@@ -11,12 +11,15 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import com.mapbox.common.Cancelable
 import com.mapbox.geojson.Point
+import com.mapbox.maps.CameraChangedCallback
 import com.mapbox.maps.CameraOptions
 import com.mapbox.maps.CameraState
 import com.mapbox.maps.EdgeInsets
 import com.mapbox.maps.MapView
 import com.mapbox.maps.MapboxExperimental
+import com.mapbox.maps.MapboxMap
 import com.mapbox.maps.dsl.cameraOptions
+import com.mapbox.maps.extension.compose.internal.StateDispatcher
 import com.mapbox.maps.logW
 import com.mapbox.maps.plugin.animation.CameraAnimationsPlugin
 import com.mapbox.maps.plugin.animation.MapAnimationOptions
@@ -24,15 +27,23 @@ import com.mapbox.maps.plugin.animation.camera
 import com.mapbox.maps.plugin.viewport.CompletionListener
 import com.mapbox.maps.plugin.viewport.ViewportPlugin
 import com.mapbox.maps.plugin.viewport.ViewportStatus
+import com.mapbox.maps.plugin.viewport.ViewportStatusObserver
 import com.mapbox.maps.plugin.viewport.data.DefaultViewportTransitionOptions
 import com.mapbox.maps.plugin.viewport.data.FollowPuckViewportStateOptions
 import com.mapbox.maps.plugin.viewport.data.OverviewViewportStateOptions
 import com.mapbox.maps.plugin.viewport.data.ViewportStatusChangeReason
+import com.mapbox.maps.plugin.viewport.state.FollowPuckViewportState
+import com.mapbox.maps.plugin.viewport.state.OverviewViewportState
 import com.mapbox.maps.plugin.viewport.state.ViewportState
 import com.mapbox.maps.plugin.viewport.state.ViewportStateDataObserver
 import com.mapbox.maps.plugin.viewport.transition.ViewportTransition
 import com.mapbox.maps.plugin.viewport.viewport
 import com.mapbox.maps.toCameraOptions
+import kotlinx.coroutines.CoroutineName
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.launch
 
 /**
  * Create and [rememberSaveable] a [MapViewportState] using [MapViewportState.Saver].
@@ -53,55 +64,107 @@ public inline fun rememberMapViewportState(
  * A [MapViewportState] may only be used by a single [MapboxMap] composable at a time
  * as it reflects instance state for a single view of a map.
  *
- * @param cameraState the initial camera position
+ * @param initialCameraState the initial camera position
  */
 @MapboxExperimental
 public class MapViewportState(
-  cameraState: CameraState = INIT_CAMERA_STATE
+  initialCameraState: CameraState = INIT_CAMERA_STATE
 ) {
   private var controller: MapView? = null
 
-  // cache the last pending operation when the operation is ran before map is attached.
-  private var lastPendingOperation: ((MapView) -> Unit)? = null
+  private val coroutineScope =
+    CoroutineScope(Dispatchers.Main.immediate + SupervisorJob() + CoroutineName("MapViewportStateScope"))
+  private var cameraChangedCancelable: Cancelable? = null
 
-  internal fun setMap(mapView: MapView?) {
-    this.controller = mapView
-    if (cameraState !== INIT_CAMERA_STATE) {
-      setCameraOptions(cameraState.toCameraOptions())
-    }
-    // invoke the last pending operation if map is valid
-    // or clear the last pending operation if the map is null
-    mapView?.let {
-      lastPendingOperation?.invoke(it)
-    } ?: kotlin.run {
-      lastPendingOperation = null
+  private val cameraChangedCallback = CameraChangedCallback {
+    cameraState = it.cameraState
+  }
+
+  private val viewportStatusObserver = ViewportStatusObserver { _, to, reason ->
+    mapViewportStatus = to
+    mapViewportStatusChangedReason = reason
+  }
+
+  private var stateDispatcher = StateDispatcher()
+
+  /**
+   * The current [CameraState] of the map or null if this [MapViewportState] is not yet associated
+   * with a map.
+   */
+  @MapboxExperimental
+  public var cameraState: CameraState? by mutableStateOf(null)
+    internal set
+
+  /**
+   * The reason why the [ViewportStatus] has been changed or null if this [MapViewportState] is not
+   * yet associated with a map.
+   */
+  @MapboxExperimental
+  public var mapViewportStatusChangedReason: ViewportStatusChangeReason? by mutableStateOf(null)
+    internal set
+
+  /**
+   * The current [ViewportStatus] that represents the status of the viewport or null if this
+   * [MapViewportState] is not yet associated with a map.
+   *
+   * It could be either a [ViewportStatus.State], [ViewportStatus.Transition] or
+   * [ViewportStatus.Idle].
+   */
+  @MapboxExperimental
+  public var mapViewportStatus: ViewportStatus? by mutableStateOf(null)
+    internal set
+
+  init {
+    // If the user has provided a specific camera state then we need to enqueue it
+    if (initialCameraState != INIT_CAMERA_STATE) {
+      setCameraOptions(initialCameraState.toCameraOptions())
     }
   }
 
-  /**
-   * The current [CameraState] of the map.
-   */
-  @MapboxExperimental
-  public var cameraState: CameraState by mutableStateOf(cameraState)
-    internal set
+  internal fun setMap(mapView: MapView?) {
+    // If the MapViewportState is set to another map with different reference, we log a warning message that we don't support multiple map instances.
+    if (controller != null && mapView != null && mapView !== this.controller) {
+      logW(
+        TAG,
+        "The map viewport state should not be used across multiple map instances! The previous map instance will lose updates."
+      )
+    }
 
-  /**
-   * The reason why the [ViewportStatus] has been changed.
-   */
-  @MapboxExperimental
-  public var mapViewportStatusChangedReason: ViewportStatusChangeReason by mutableStateOf(
-    ViewportStatusChangeReason.IDLE_REQUESTED
-  )
-    internal set
+    // Setting same map instance will be a no-op
+    if (mapView === this.controller) {
+      return
+    }
 
-  /**
-   * The current [ViewportStatus] that represents the status of the viewport.
-   *
-   * It could be either a [ViewportStatus.State], [ViewportStatus.Transition] or [ViewportStatus.Idle].
-   */
-  @MapboxExperimental
-  public var mapViewportStatus: ViewportStatus by mutableStateOf(ViewportStatus.Idle)
-    internal set
+    // clear up previous observers
+    controller?.viewport?.removeStatusObserver(viewportStatusObserver)
+    cameraChangedCancelable?.cancel()
+
+    // update controller
+    this.controller = mapView
+
+    // Resume or pause the dispatching based on current validness of mapview
+    if (mapView != null) {
+      // Manually retrieve the current camera state, so we don't need to wait for the first callback.
+      cameraState = mapView.mapboxMap.cameraState
+      // start observing the camera updates from the new mapview
+      cameraChangedCancelable = mapView.mapboxMap.subscribeCameraChanged(cameraChangedCallback)
+
+      // Manually retrieve the current viewport state, so we don't need to wait for the first callback.
+      val viewport = mapView.viewport
+      mapViewportStatus = viewport.status
+      // TODO: we don't expose the reason as public API in viewport plugin so for now we reset to null
+      mapViewportStatusChangedReason = null
+      viewport.addStatusObserver(viewportStatusObserver)
+
+      stateDispatcher.resumeDispatching()
+    } else {
+      // We lost the mapview so let's reset the current state to unknown
+      cameraState = null
+      mapViewportStatus = null
+      mapViewportStatusChangedReason = null
+      stateDispatcher.pauseDispatching()
+    }
+  }
 
   /**
    * Move the camera instantaneously as specified by [cameraOptions]. Any camera animation in progress
@@ -112,7 +175,7 @@ public class MapViewportState(
   @UiThread
   @MapboxExperimental
   public fun setCameraOptions(cameraOptions: CameraOptions) {
-    tryInvokingOperation { mapView ->
+    dispatchMapViewOperation { mapView ->
       mapView.apply {
         viewport.transitionTo(
           CameraViewportState(cameraOptions) {
@@ -137,7 +200,8 @@ public class MapViewportState(
   }
 
   /**
-   * A utility function to get the default [CameraOptions] defined in the style.
+   * A utility function to get the default [CameraOptions] defined in the style or null if this
+   * [MapViewportState] is not yet associated with a map.
    */
   @MapboxExperimental
   public val styleDefaultCameraOptions: CameraOptions?
@@ -156,7 +220,7 @@ public class MapViewportState(
     animationOptions: MapAnimationOptions? = null,
     completionListener: CompletionListener? = null
   ) {
-    tryInvokingOperation { mapView ->
+    dispatchMapViewOperation { mapView ->
       mapView.apply {
         viewport.transitionTo(
           targetState = CameraViewportState(cameraOptions) {
@@ -196,7 +260,7 @@ public class MapViewportState(
     animationOptions: MapAnimationOptions? = null,
     completionListener: CompletionListener? = null
   ) {
-    tryInvokingOperation { mapView ->
+    dispatchMapViewOperation { mapView ->
       mapView.apply {
         viewport.transitionTo(
           targetState = CameraViewportState(cameraOptions) {
@@ -227,7 +291,7 @@ public class MapViewportState(
       .build(),
     completionListener: CompletionListener? = null
   ) {
-    tryInvokingOperation { mapView ->
+    dispatchMapViewOperation { mapView ->
       mapView.viewport.apply {
         transitionTo(
           targetState = makeOverviewViewportState(overviewViewportStateOptions),
@@ -247,12 +311,13 @@ public class MapViewportState(
    */
   @MapboxExperimental
   public fun transitionToFollowPuckState(
-    followPuckViewportStateOptions: FollowPuckViewportStateOptions = FollowPuckViewportStateOptions.Builder().build(),
+    followPuckViewportStateOptions: FollowPuckViewportStateOptions = FollowPuckViewportStateOptions.Builder()
+      .build(),
     defaultTransitionOptions: DefaultViewportTransitionOptions = DefaultViewportTransitionOptions.Builder()
       .build(),
     completionListener: CompletionListener? = null
   ) {
-    tryInvokingOperation { mapView ->
+    dispatchMapViewOperation { mapView ->
       mapView.viewport.apply {
         transitionTo(
           targetState = makeFollowPuckViewportState(followPuckViewportStateOptions),
@@ -270,22 +335,17 @@ public class MapViewportState(
    */
   @MapboxExperimental
   public fun idle() {
-    tryInvokingOperation { mapView ->
+    dispatchMapViewOperation { mapView ->
       mapView.viewport.idle()
     }
   }
 
-  private fun tryInvokingOperation(operation: (MapView) -> Unit) {
-    with(controller) {
-      if (this == null) {
-        logW(
-          TAG,
-          "tryInvokingOperation invoked when Map is not set, added the operation as pending operation."
-        )
-        lastPendingOperation = operation
-      } else {
-        operation.invoke(this)
+  private fun dispatchMapViewOperation(operation: (MapView) -> Unit) {
+    coroutineScope.launch {
+      val block: suspend CoroutineScope.() -> Unit = {
+        controller!!.apply(operation)
       }
+      stateDispatcher.dispatch(block = block)
     }
   }